fix: enforce logical consistency for profiles in seed service (#545)
